feat(api): add callBackendApi and callNigeriaApi
Add the `callBackendApi` and `callNigeriaApi` functions to the API module. These functions are responsible for making API calls to the backend servers. The `callBackendApi` function is used to make calls to the SRM API, while the `callNigeriaApi` function is used to make calls to the Nigeria States, Towns, and LGAs API. These functions are created using the `callApi` utility from the `@zayne-labs/callapi` package.